refactor(types): move interfaces to a separate file
