+++ src/types/components/serviceSection.ts
@@ -1,0 +1,18 @@
+export interface IBacklink {
+  id: string
+  text: string
+}
+
+export interface IServiceContent {
+  id: string
+  title: string
+  description: string
+  backlinks: IBacklink[]
+  image: string
+  link: string
+  category: string
+}
+
+export type ServicesContentProps = {
+  [key: string]: IServiceContent
+}
