+++ src/types/ui/tabs.ts
@@ -1,0 +1,11 @@
+export interface ITabProperties {
+  key: string
+  label: string
+  children: React.ReactNode
+}
+
+export interface IVerticalTabs {
+  key: string
+  label: string
+  children: React.ReactNode
+}
